refactor: refactor implementation of project popover
